requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
alembic
passlib[bcrypt]
python-jose[cryptography]
pydantic
pydantic[email]
python-multipart
bcrypt==4.0.1
passlib==1.7.4